% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_download_dir.r
\name{get_download_dir}
\alias{get_download_dir}
\title{Get Default Dewey Download Directory}
\usage{
get_download_dir(create = TRUE)
}
\arguments{
\item{create}{Logical. If TRUE, creates the directory if it doesn't exist.
Default is TRUE.}
}
\value{
A character string containing the absolute path to the download directory
}
\description{
Returns the absolute path to the default download directory for Dewey data,
which is a "dewey-downloads" folder in the current working directory.
}
\examples{
# Get default download path (creates if doesn't exist)
get_download_dir()

# Get path without creating directory
get_download_dir(create = FALSE)

}
